requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0